requirements.txt
faster-whisper>=1.0.0
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
requests>=2.28.0
pydantic>=2.0
sherpa-onnx>=1.10.0
soundfile>=0.12.1
numpy>=1.25.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.25
pymysql>=1.1.0
python-multipart
transliterate
